Extract Preset class into its own file
