+++ OpenTabletDriver.Desktop/Preset.cs
@@ -1,0 +1,21 @@
+using Newtonsoft.Json;
+
+namespace OpenTabletDriver.Desktop
+{
+    public class Preset
+    {
+        public Preset(string name, Settings settings)
+        {
+            Name = name;
+            Settings = settings;
+        }
+
+        public string Name { get; }
+        private Settings Settings;
+
+        public Settings GetSettings()
+        {
+            return JsonConvert.DeserializeObject<Settings>(JsonConvert.SerializeObject(Settings));
+        }
+    }
+}
